Friday 7/11 Commit 1 - Class/File Organization and Renaming (Scanners and Abstracts)
- Moved abstract methods to abstract_controller.py and renamed it to just abstract.py. Made CameraThread a nested class in AbstractScanner.

- Removed outdated docstrings.

- Renamed Scanner() in scanner.py to ItemScanner.

- Created file_exports.py and qr_generator.py for better organization and modularity. Moved all file export methods to file_exports.py.

- Fixed imports and added new files and classes to __all__ in __init__.py files.
