+++ stock_manager/controllers/__init__.py
@@ -1,23 +1,20 @@
-"""
-Controllers package for the Stock Management Application.
+"""Controllers package for the Stock Management Application."""
 
-Exposes the main View and Scanner controllers for the application's UI logic.
-"""
-
-from .abstract_controller import AbstractController
+from .abstract import AbstractController, AbstractScanner
 from .add import Add
 from .edit import Edit
 from .export import Export
 from .finish import Finish
 from .remove import Remove
-from .scanner import Login, Scanner
+from .scanner import ItemScanner, Login
 from .view import View
 
 __all__ = [
 	'AbstractController',
+	'AbstractScanner',
 	'Login',
 	'View',
-	'Scanner',
+	'ItemScanner',
 	'Finish',
 	'Export',
 	'Add',
