+++ stock_manager/utils/__init__.py
@@ -6,9 +6,13 @@ Exposes Logger and DBUtils classes for logging and Google Sheets access.
 
 from .constants import *
 from .database import DBUtils
+from .file_exports import FileExports
 from .logger import Logger
+from .qr_generator import QRGenerator
 
 __all__ = [
 			  'Logger',
-			  'DBUtils'
+			  'DBUtils',
+			  'FileExports',
+			  'QRGenerator'
 		  ] + [name for name in dir(constants)]
