+++ stock_manager/utils/qr_generator.py
@@ -1,0 +1,2 @@
+class QRGenerator:
+	pass
